requirements.txt
django==1.11.*
psycopg2-binary==2.8.*
python-dotenv==0.15.0
environs==9.3.1